cli: add windows amd64 build target
