+++ cli/BUILD.bazel
@@ -37,6 +37,7 @@ go_binary(
         "darwin_arm64",
         "linux_amd64",
         "linux_arm64",
+        "windows_amd64",
     ]
     for edition in [
         "oss",
