+++ go.mod
@@ -70,7 +70,7 @@ require (
 	github.com/docker/docker v23.0.3+incompatible
 	github.com/edgelesssys/constellation/v2/operators/constellation-node-operator/v2/api v0.0.0
 	github.com/edgelesssys/go-azguestattestation v0.0.0-20230303085714-62ede861d33f
-	github.com/edgelesssys/go-tdx-qpl v0.0.0-20230307140231-bb361f158928
+	github.com/edgelesssys/go-tdx-qpl v0.0.0-20230530085549-fd2878a4dead
 	github.com/fsnotify/fsnotify v1.6.0
 	github.com/go-playground/locales v0.14.1
 	github.com/go-playground/universal-translator v0.18.1
